修改double 转换问题和UIbug

+++ CRTBTunnelMonitor/src/com/crtb/tunnelmonitor/widget/CrtbRecordSubsidenceAdapter.java
@@ -45,7 +45,7 @@ public class CrtbRecordSubsidenceAdapter extends CrtbEntityAdapter<RawSheetIndex
 		}
 		
 		holder.chainage.setText(DateUtils.toDateString(item.getCreateTime(),DateUtils.DATE_TIME_FORMAT));
-		holder.excavation.setText(CrtbUtils.formatSectionName(item.getPrefix(), (float)item.getFACEDK()));
+		holder.excavation.setText(CrtbUtils.formatSectionName(item.getPrefix(), item.getFACEDK()));
 		
 		return convertView;
 	}
