+++ CRTBTunnelMonitor/src/com/crtb/tunnelmonitor/widget/CrtbTestRecordSubsidenceAdapter.java
@@ -78,7 +78,7 @@ public class CrtbTestRecordSubsidenceAdapter extends CrtbEntityAdapter<RawSheetI
 		}
 		
 		holder.recordNo.setText(String.valueOf(item.getID()));
-		holder.recordName.setText(CrtbUtils.formatSectionName(item.getPrefix(),(float)item.getFACEDK()));
+		holder.recordName.setText(CrtbUtils.formatSectionName(item.getPrefix(),item.getFACEDK()));
 		
 		if(!item.isChecked()){
 			holder.status.setBackgroundResource(R.drawable.no);
